+++ v2/connection/connection_http2.go
@@ -21,8 +21,6 @@
 package connection
 
 import (
-	"net/http"
-
 	"golang.org/x/net/http2"
 )
 
@@ -34,10 +32,10 @@ type Http2Configuration struct {
 
 	ArangoDBConfig ArangoDBConfiguration
 
-	Transport http.RoundTripper
+	Transport *http2.Transport
 }
 
-func (h Http2Configuration) getTransport() http.RoundTripper {
+func (h Http2Configuration) getTransport() *http2.Transport {
 	if h.Transport != nil {
 		return h.Transport
 	}